refactor: duplicate codes are put together
