+++ crates/erg_compiler/lower.rs
@@ -3,16 +3,13 @@
 //! ASTLowerer(ASTからHIRへの変換器)を実装
 use std::mem;
 
+use erg_common::config::{ErgConfig, ErgMode};
 use erg_common::dict;
 use erg_common::error::{Location, MultiErrorDisplay};
 use erg_common::set;
 use erg_common::set::Set;
 use erg_common::traits::{Locational, NoTypeDisplay, Runnable, Stream};
 use erg_common::vis::Visibility;
-use erg_common::{
-    config::{ErgConfig, ErgMode},
-    traits::BlockKind,
-};
 use erg_common::{fmt_option, fn_name, log, option_enum_unwrap, switch_lang, Str};
 
 use erg_parser::ast;
@@ -122,43 +119,6 @@ impl Runnable for ASTLowerer {
             .map_err(|artifact| artifact.errors)?;
         artifact.warns.fmt_all_stderr();
         Ok(format!("{}", artifact.object))
-    }
-
-    #[inline]
-    fn expect_block(&self, src: &str) -> BlockKind {
-        let multi_line_str = "\"\"\"";
-        if src.contains(multi_line_str) && src.rfind(multi_line_str) == src.find(multi_line_str) {
-            return BlockKind::MultiLineStr;
-        }
-        if src.trim_start().starts_with('@') {
-            return BlockKind::AtMark;
-        }
-        if src.ends_with("do!:") && !src.starts_with("do!:") {
-            return BlockKind::Lambda;
-        }
-        if src.ends_with("do:") && !src.starts_with("do:") {
-            return BlockKind::Lambda;
-        }
-        if src.ends_with(':') && !src.starts_with(':') {
-            return BlockKind::Lambda;
-        }
-        if src.ends_with('=') && !src.starts_with('=') {
-            return BlockKind::Assignment;
-        }
-        if src.ends_with('.') && !src.starts_with('.') {
-            return BlockKind::ClassPub;
-        }
-        if src.ends_with("::") && !src.starts_with("::") {
-            return BlockKind::ClassPriv;
-        }
-        if src.ends_with("=>") && !src.starts_with("=>") {
-            return BlockKind::Lambda;
-        }
-        if src.ends_with("->") && !src.starts_with("->") {
-            return BlockKind::Lambda;
-        }
-
-        BlockKind::None
     }
 }
