+++ crates/erg_compiler/transpile.rs
@@ -4,7 +4,7 @@ use std::io::Write;
 use erg_common::config::ErgConfig;
 use erg_common::error::MultiErrorDisplay;
 use erg_common::log;
-use erg_common::traits::{BlockKind, Runnable, Stream};
+use erg_common::traits::{Runnable, Stream};
 use erg_common::Str;
 
 use erg_parser::ast::{ParamPattern, TypeSpec, VarName};
@@ -163,43 +163,6 @@ impl Runnable for Transpiler {
         })?;
         artifact.warns.fmt_all_stderr();
         Ok(artifact.object.code)
-    }
-
-    #[inline]
-    fn expect_block(&self, src: &str) -> BlockKind {
-        let multi_line_str = "\"\"\"";
-        if src.contains(multi_line_str) && src.rfind(multi_line_str) == src.find(multi_line_str) {
-            return BlockKind::MultiLineStr;
-        }
-        if src.trim_start().starts_with('@') {
-            return BlockKind::AtMark;
-        }
-        if src.ends_with("do!:") && !src.starts_with("do!:") {
-            return BlockKind::Lambda;
-        }
-        if src.ends_with("do:") && !src.starts_with("do:") {
-            return BlockKind::Lambda;
-        }
-        if src.ends_with(':') && !src.starts_with(':') {
-            return BlockKind::Lambda;
-        }
-        if src.ends_with('=') && !src.starts_with('=') {
-            return BlockKind::Assignment;
-        }
-        if src.ends_with('.') && !src.starts_with('.') {
-            return BlockKind::ClassPub;
-        }
-        if src.ends_with("::") && !src.starts_with("::") {
-            return BlockKind::ClassPriv;
-        }
-        if src.ends_with("=>") && !src.starts_with("=>") {
-            return BlockKind::Lambda;
-        }
-        if src.ends_with("->") && !src.starts_with("->") {
-            return BlockKind::Lambda;
-        }
-
-        BlockKind::None
     }
 }
