+++ crates/erg_parser/lex.rs
@@ -7,7 +7,7 @@ use erg_common::cache::CacheSet;
 use erg_common::config::ErgConfig;
 use erg_common::config::Input;
 use erg_common::traits::DequeStream;
-use erg_common::traits::{BlockKind, Locational, Runnable, Stream};
+use erg_common::traits::{Locational, Runnable, Stream};
 use erg_common::{debug_power_assert, fn_name_full, normalize_newline, switch_lang};
 
 use crate::error::{LexError, LexErrors, LexResult, LexerRunnerError, LexerRunnerErrors};
@@ -87,43 +87,6 @@ impl Runnable for LexerRunner {
                 .map_err(|errs| LexerRunnerErrors::convert(self.input(), errs))?
                 .to_string())
         }
-    }
-
-    #[inline]
-    fn expect_block(&self, src: &str) -> BlockKind {
-        let multi_line_str = "\"\"\"";
-        if src.contains(multi_line_str) && src.rfind(multi_line_str) == src.find(multi_line_str) {
-            return BlockKind::MultiLineStr;
-        }
-        if src.trim_start().starts_with('@') {
-            return BlockKind::AtMark;
-        }
-        if src.ends_with("do!:") && !src.starts_with("do!:") {
-            return BlockKind::Lambda;
-        }
-        if src.ends_with("do:") && !src.starts_with("do:") {
-            return BlockKind::Lambda;
-        }
-        if src.ends_with(':') && !src.starts_with(':') {
-            return BlockKind::Lambda;
-        }
-        if src.ends_with('=') && !src.starts_with('=') {
-            return BlockKind::Assignment;
-        }
-        if src.ends_with('.') && !src.starts_with('.') {
-            return BlockKind::ClassPub;
-        }
-        if src.ends_with("::") && !src.starts_with("::") {
-            return BlockKind::ClassPriv;
-        }
-        if src.ends_with("=>") && !src.starts_with("=>") {
-            return BlockKind::Lambda;
-        }
-        if src.ends_with("->") && !src.starts_with("->") {
-            return BlockKind::Lambda;
-        }
-
-        BlockKind::None
     }
 }
